pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "magnetic-levitation-sim"
version = "0.1.0"
description = "Research-grade magnetic levitation simulation and control stack for a spherical robot in a cube"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "matplotlib>=3.8",
  "numpy>=1.26",
  "PySide6>=6.7",
  "pyqtgraph>=0.13",
  "PyYAML>=6.0.1",
  "scipy>=1.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

